+++ libsolidity/codegen/ir/IRGeneratorForStatements.cpp
@@ -692,16 +692,16 @@ void IRGeneratorForStatements::endVisit(FunctionCall const& _functionCall)
 				joinHumanReadable(args) <<
 				")\n";
 		else
+		{
+			YulArity arity = YulArity::fromType(*functionType);
 			define(_functionCall) <<
-				// NOTE: internalDispatch() takes care of adding the function to function generation queue
-				m_context.internalDispatch(
-					TupleType(functionType->parameterTypes()).sizeOnStack(),
-					TupleType(functionType->returnParameterTypes()).sizeOnStack()
-				) <<
+				// NOTE: generateInternalDispatchFunction() takes care of adding the function to function generation queue
+				m_context.generateInternalDispatchFunction(arity) <<
 				"(" <<
 				IRVariable(_functionCall.expression()).part("functionIdentifier").name() <<
 				joinHumanReadablePrefixed(args) <<
 				")\n";
+		}
 		break;
 	}
 	case FunctionType::Kind::External: