+++ libsolidity/formal/SymbolicState.cpp
@@ -20,6 +20,7 @@
 #include <libsolidity/formal/EncodingContext.h>
 
 using namespace std;
+using namespace solidity;
 using namespace solidity::frontend::smt;
 
 SymbolicState::SymbolicState(EncodingContext& _context):
@@ -35,22 +36,22 @@ void SymbolicState::reset()
 
 // Blockchain
 
-Expression SymbolicState::thisAddress()
+smtutil::Expression SymbolicState::thisAddress()
 {
 	return m_thisAddress.currentValue();
 }
 
-Expression SymbolicState::balance()
+smtutil::Expression SymbolicState::balance()
 {
 	return balance(m_thisAddress.currentValue());
 }
 
-Expression SymbolicState::balance(Expression _address)
+smtutil::Expression SymbolicState::balance(smtutil::Expression _address)
 {
-	return Expression::select(m_balances.elements(), move(_address));
+	return smtutil::Expression::select(m_balances.elements(), move(_address));
 }
 
-void SymbolicState::transfer(Expression _from, Expression _to, Expression _value)
+void SymbolicState::transfer(smtutil::Expression _from, smtutil::Expression _to, smtutil::Expression _value)
 {
 	unsigned indexBefore = m_balances.index();
 	addBalance(_from, 0 - _value);
@@ -59,7 +60,7 @@ void SymbolicState::transfer(Expression _from, Expression _to, Expression _value
 	solAssert(indexAfter > indexBefore, "");
 	m_balances.increaseIndex();
 	/// Do not apply the transfer operation if _from == _to.
-	auto newBalances = Expression::ite(
+	auto newBalances = smtutil::Expression::ite(
 		move(_from) == move(_to),
 		m_balances.valueAtIndex(indexBefore),
 		m_balances.valueAtIndex(indexAfter)
@@ -69,9 +70,9 @@ void SymbolicState::transfer(Expression _from, Expression _to, Expression _value
 
 /// Private helpers.
 
-void SymbolicState::addBalance(Expression _address, Expression _value)
+void SymbolicState::addBalance(smtutil::Expression _address, smtutil::Expression _value)
 {
-	auto newBalances = Expression::store(
+	auto newBalances = smtutil::Expression::store(
 		m_balances.elements(),
 		_address,
 		balance(_address) + move(_value)