+++ test/libsolidity/SMTCheckerTest.cpp
@@ -30,13 +30,13 @@ SMTCheckerTest::SMTCheckerTest(string const& _filename, langutil::EVMVersion _ev
 {
 	auto const& choice = m_reader.stringSetting("SMTSolvers", "any");
 	if (choice == "any")
-		m_enabledSolvers = smt::SMTSolverChoice::All();
+		m_enabledSolvers = smtutil::SMTSolverChoice::All();
 	else if (choice == "z3")
-		m_enabledSolvers = smt::SMTSolverChoice::Z3();
+		m_enabledSolvers = smtutil::SMTSolverChoice::Z3();
 	else if (choice == "cvc4")
-		m_enabledSolvers = smt::SMTSolverChoice::CVC4();
+		m_enabledSolvers = smtutil::SMTSolverChoice::CVC4();
 	else if (choice == "none")
-		m_enabledSolvers = smt::SMTSolverChoice::None();
+		m_enabledSolvers = smtutil::SMTSolverChoice::None();
 	else
 		BOOST_THROW_EXCEPTION(runtime_error("Invalid SMT solver choice."));