+++ test/libsolidity/syntaxTests/natspec/docstring_private_state_variable.sol
@@ -1,0 +1,7 @@
+contract C {
+    /// @notice example of notice
+    /// @dev example of dev
+    uint private state;
+}
+// ----
+// Warning: (17-74): Documentation tag on non-public state variables will be disallowed in 0.7.0. You will need to use the @dev tag explicitly.